requirements.txt
gunicorn==19.6.0
python-box==3.0.1
Jinja2==2.9.5
requests==2.10.0
Flask==0.12
psycopg2==2.6.2
requests-cache==0.4.13
mwclient==0.8.6
